+++ Sources/Paywall/Paywall/Events/Internal Tracking/Trackable Events/UserInitiatedEvents.swift
@@ -18,7 +18,14 @@ enum UserInitiatedEvent {
     let canImplicitlyTriggerPaywall: Bool
     var customParameters: [String: Any] = [:]
   }
-  
+
+  struct DefaultPaywall: TrackableUserInitiatedEvent {
+    let rawName = "$present"
+    let superwallParameters: [String: Any] = [:]
+    let canImplicitlyTriggerPaywall = false
+    let customParameters: [String: Any] = [:]
+  }
+
   // MARK: - To be deprecated/deleted
   struct PushNotification: TrackableUserInitiatedEvent {
     enum State {